lambda/requirements.txt
ask-sdk-core==1.11.0
beautifulsoup4==4.9.3
requests==2.22.0
psycopg2==2.8.6
PyYAML==5.3.1